adding . for xpath locators moving page objects to a new package
