+++ src/test/java/test/pageobjects/TeachersPage.java
@@ -1,9 +1,12 @@
-package test;
+package test.pageobjects;
 
+import org.openqa.selenium.By;
 import org.openqa.selenium.Keys;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
+
+import test.pageobjects.base.PageObject;
 
 public class TeachersPage extends PageObject {
 
@@ -14,27 +17,38 @@ public class TeachersPage extends PageObject {
 
 	@FindBy(className = "MuiFab-label")
 	private WebElement addIcon;
-	@FindBy(xpath = "//input[@name='name']")
+	
+	@FindBy(xpath = ".//input[@name='name']")
 	private WebElement name;
-	@FindBy(xpath = "//input[@name='surname']")
+	
+	@FindBy(xpath = ".//input[@name='surname']")
 	private WebElement surname;
-	@FindBy(xpath = "//input[@name='email']")
+	
+	@FindBy(xpath = ".//input[@name='email']")
 	private WebElement email;
-	@FindBy(xpath = "//div[@class='makeStyles-actions-13']/button[@class='MuiButtonBase-root MuiButton-root MuiButton-contained MuiButton-containedPrimary']")
+	
+	@FindBy(xpath = ".//div[@class='makeStyles-actions-13']/button[@class='MuiButtonBase-root MuiButton-root MuiButton-contained MuiButton-containedPrimary']")
 	private WebElement saveButton;
+	
 	@FindBy(className = "MuiDataGrid-mainGridContainer")
 	private WebElement anyData;
-	@FindBy(xpath = "//div[@class='makeStyles-actions-13']/button[@class='MuiButtonBase-root MuiButton-root MuiButton-contained MuiButton-containedSecondary']")
+	
+	@FindBy(xpath = ".//div[@class='makeStyles-actions-13']/button[@class='MuiButtonBase-root MuiButton-root MuiButton-contained MuiButton-containedSecondary']")
 	private WebElement deleteButton;
+	
 	@FindBy(className = "MuiButton-label")
 	private WebElement toggleCourses;
-	@FindBy(xpath = "//div[not(@*)]/button[@class='MuiButtonBase-root MuiButton-root MuiButton-contained MuiButton-containedPrimary']")
-	private WebElement assignNewCourseButton;	
-	@FindBy(xpath = "//input[@name='developerCourseId']")
+	
+	@FindBy(xpath = ".//div[not(@*)]/button[@class='MuiButtonBase-root MuiButton-root MuiButton-contained MuiButton-containedPrimary']")
+	private WebElement assignNewCourseButton;
+	
+	@FindBy(xpath = ".//input[@name='developerCourseId']")
 	private WebElement course;
-	@FindBy(xpath = "//div[@class='MuiPaper-root MuiMenu-paper MuiPopover-paper MuiPaper-elevation8 MuiPaper-rounded']")
+	
+	@FindBy(xpath = ".//div[@class='MuiPaper-root MuiMenu-paper MuiPopover-paper MuiPaper-elevation8 MuiPaper-rounded']")
 	private WebElement courseFromList;
-	@FindBy(xpath = "//div[@class='makeStyles-actions-16']/button[@class='MuiButtonBase-root MuiButton-root MuiButton-contained MuiButton-containedPrimary']")
+	
+	@FindBy(xpath = ".//div[@class='makeStyles-actions-16']/button[@class='MuiButtonBase-root MuiButton-root MuiButton-contained MuiButton-containedPrimary']")
 	private WebElement saveCourseButton;
 
 	public void enterName(String name) {
